refactor: improve class attendance data handling and fix API endpoint
- Update `AxisAttendanceClient` with the correct session attendance endpoint.
- Change `startTimeMs` and `endTimeMs` types from `int` to `long` to prevent overflow.
- Modify `OfferingResponse` to use `String` for time fields.
- Refactor `ClassAttendanceFragment` and `ClassAttendanceAdapter` to use a list-based data structure and notify adapters of changes.
- Initialize `MutableLiveData` in `ClassAttendanceViewModel` with empty default values.
- Add safety checks for session retrieval and update summary text formatting.

+++ app/src/main/java/com/example/faculty_app/mainapp/attendance/data/local/models/sessions/SessionModel.java
@@ -9,15 +9,15 @@ public class SessionModel {
     public String status;
     @NonNull
     public String date;
-    public int startTimeMs;
-    public int getStartTimeMs;
+    public long startTimeMs;
+    public long getStartTimeMs;
 
     public SessionModel(int id,
                         int classOffering,
                         @NonNull String status,
                         @NonNull String date,
-                        int startTimeMs,
-                        int getStartTimeMs) {
+                        long startTimeMs,
+                        long getStartTimeMs) {
         this.id = id;
         this.classOffering = classOffering;
         this.status = status;
